schulze-evaluate: require stable input sorting of candidates
schulze_evaluate gives a stable result for arbitary sorting of candidates,
votes and equal candidates inside the votes. But the result may differ in
the order of equal candidates in the provided result.

To obtain an identical sorting here, we require a stable sorting of
candidates. Since we do not know after which criteria the candidates
shall be sorted, we left this to the user.